requirements.txt
fastapi
numpy
pandas
pytest
scipy
scikit_learn
setuptools
requests
pytest-asyncio
aiohttp
httpx
uvicorn
gunicorn
PyYAML
-e .